+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. This change
 ## [Unreleased]
 ### Added
 - Creation of `Featurelike` protocol
+- Initial support for reading shapefiles
+- Initial support for specs, generators, and generative testing of `Featurelike`
 
 [Unreleased]: https://github.com/willcohen/ovid/compare/0.1.1...HEAD
 [0.1.1]: https://github.com/willcohen/ovid/compare/0.1.0...0.1.1